Testing: Resolve makedep paths
This patch resolves some issues relating to finding the path to makedep
in both .testing and more generalized autoconf builds.

An explicit autoconf test for makedep has been added, with a default
path which includes the `ac` directory relative to `deps`.

The .testing directory, which does not lie within `ac`, instead modifies
the PATH to allow autoconf to find makedep.

The absolute path is determined and substituted into the appropriate
Makefile.in template.

Some redundant operations in .testing/Makefile have been removed, but I
suspect there are even more.  Much of the structure required to support
mkmf and list_paths is probably no longer needed.