Remove roundoff mitigation from PLM_slope_cw()
We have code to mitigate round-off troubles with the PLM schemes which
deliberately underestimate slopes in the hope that when edge values
are reconstructed that they don't have unwanted overshoots. This
is not robust and for future schemes we will be transitioning to an
edge value description of all schemes (even though redundant for some)
which will be robust. Since PLM_slope_cw() is unused, I am preempting
having to duplicate this function and commandeering it for a one of the
said future schemes.