Improved documentation of PLM APIs
The PLM_functions module APIs were documented but did not describe
the details enough that we could easily misinterpret the functions (and
have been misusing them as a result). Several choices about the
details matter and I will be adding alternative versions with different
choices to use in forthcoming reconstructions.

- Added descriptions of algorithms in the API descriptions
= Added references where missing