Antarctic hole treatment, and miscellaneous fixes
-Added ice-sheet mass to the MOM stocks, including ice-sheet mass in
the hole in the MOM grid at the S. pole
Hole mass is calculated by: integrating dh_s (the
thickness change due to adot, which is not necessarily the same as
adot*dt if some cells enforce constant thickness) over the regular
MOM-IS grid; adding ice-flow in/out of the hole; and subtracting
either:
(1) adot*dt integrated over the ice sheet using the modified
INPUT_lndXIS/h_grid.nc cell areas, or
(2) adot*dt integrated over the ice sheet within the land
model.
The modified grid has extended edges of the southernmost cells
near the S. Pole to nearly cover the S. Pole hole (e.g. -89.99999),
with corresponding modified cell areas. This grid is also used for the
land-to-ice exchange of adot.
-New MOM_IS parameter: SET_GRID_HOLE_AREA = (default false) Set true
 to read in and use the modified h_grid for integrating adot in
 MOM6-IS.
-New MOM_IS_parameter: MASS_HOLE_FROM_LAND = (default true)
 Calculate hole mass using integrated adot from (True) the land model
 or (False) using the modified MOM grid areas
-New MOM_IS parameter: DEBUG_MASS_HOLE = (default false)
 Set true to ouput the integrated adot for land and MOM ice sheet, and
 the error in hole mass increment calculated using integrated adot
 from MOM vs from land.
-Hole mass also added to the mass total reported by ice_shelf.stats
-Small fixes in MOM_ice_shelf
 -Safeguard added against ice front advection past one cell if
 advection is used for calving
 -Fixed potential division by zero error when restricting surface
 slope
 -When reading in area_shelf_h, adjust it if needed to exactly match
 MOM grid areas when h_mask == 1 or h_mask == 3