Cleanup of messages to stderr emitted during run-coverage
The GitHub action "run-coverage" has what could be misinterpreted as
unit test fails but in fact are a result of testing the testing.

To fix, I've redirected the stderr messages to stdout while testing
the helper functions in the numerical_testing_type. I also added
more descriptive info in the stdout around all this emitted stuff
to help anyone who might find themselves reading it.