Finalized mass_hole treatment
Removed the option for calculating mass of the hole at the South Pole
that uses a second grid with extended areas for cells bordering the
South Pole. Instead, the now hard-coded approach is to calculate
mass_hole as the time-integrated surface mass flux area-integrated
over the land grid minus area-integrated over the ocean grid, plus any
flux in/out of the hole. This approach was determined to be simpler
and more consistent, given that the secondary exchange grid to
interpolate surface mass flux from land to ice sheet does not need to
use extended areas near the S. Pole. Instead, this secondary exchange
grid for this interpolation is the same as used for all other
exchanges, but with a different land mask.

This commit also includes a new MOM6 parameter
calculate_mass_hole (default true), which can be set false to avoid
calculating the mass in the hole entirely. It should be set false
unless ice-sheet surface mass flux is passed to the ice sheet from the
land model; otherwise, the mass_hole calculation and stocks will be
incorrect.