Adds unit_tests for coord_zlike.F90
Coverage of coord_zlike.F90 was only 60% because a branch to an second algorithm is
only triggered by the presence of optional argument which does not happen in the GH CI.
This commit adds tests invoking build_zstar_column() with various combinations of the
optional arguments checked against known results. Many are not as expected.