Replaced build_zstar_column optional args with alternative function
Optional arguments were used to indicate the presence of an ice-shelf
and the need to clip the z-grid at the top. There were TWO optional
arguments that had to be added together, and be consistent in values.
I've replaced this with a second function with fixed API so that we
don't have confusion about providing just one of the optional args.