Extend the PGF reconstruction to allow PLM-WLS
The PLM reconstruction used within the pressure gradient force
now supports the weighted least squares approach for slope
estimation.

In a catastrophic version of seamount/z where vanished layers
slightly inflate, the regular finite volume PLM method is sensitive
to values in the vanished layers and leads to a feedback that causes
en error growth (spontaneous motion). The PLM-WLS method is insensitive
to the vanished layers and in the same test leads only to round-off
level noise in the flow.