*Update TC testing parameters for late 2025
  Updated the values of about 21 parameters (many of which are repeated across
TC test cases) used in the TC testing to test the most recent versions of code
that is selected with ANSWER_DATE flags and to avoid testing the buggy versions
of code that is regulated by _BUG flags.  This includes some changes to broaden
the range of equations of state that are being tested and to test some newer
versions.  This does change the details of the TC tests, but they should (and
do) still pass TC regression tests across code versions.